+++ src/results/table.rs
@@ -3,14 +3,15 @@ use amazon_qldb_driver::transaction::StatementResults;
 use anyhow::{anyhow, Result};
 use bigdecimal::BigDecimal;
 use comfy_table::Table;
+use ion_c_sys::timestamp::IonDateTime;
 use ion_rs::value::*;
 use ion_rs::value::{
     loader::{loader, Loader},
     owned::OwnedElement,
 };
 use ion_rs::IonType;
-use std::collections::HashSet;
 use std::convert::TryFrom;
+use std::{collections::HashSet, convert::TryInto};
 
 pub(crate) fn display_results_table(results: &StatementResults, ui: &Box<dyn Ui>) -> Result<()> {
     let loader = loader();
@@ -124,7 +125,11 @@ fn format_element_for_cell(elem: Option<&OwnedElement>) -> Result<String> {
                 Err(_) => format!("-0"),
             }
         }
-        IonType::Timestamp => Err(anyhow!("timestamps are not yet supported"))?,
+        IonType::Timestamp => {
+            let ts = elem.as_timestamp().unwrap().clone();
+            let ct: IonDateTime = ts.try_into()?;
+            ct.as_datetime().to_rfc3339()
+        }
         IonType::Symbol => elem.as_sym().unwrap().text().unwrap().to_string(),
         IonType::String => elem.as_str().unwrap().to_string(),
         IonType::Clob | IonType::Blob => {