+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [2.0.0-alpha6] - [Unreleased]
 
+### Added
+
+  * Timestamps now render in `--format table`
+
 ## [2.0.0-alpha5] - 2021-04-23
 
 ### Changed